feat: 다른 유저 정보 가져오는 로직 추가

+++ backend/choki/src/main/java/com/yeojiphap/choki/domain/user/domain/User.java
@@ -35,7 +35,7 @@ public class User {
 
     private int pastLevel;
 
-    private int mainAnimal;
+    private Long mainAnimal;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "family_id")
